size_t as type for local mem size is not portable
* on 64bit platforms size_t is 64bit => platform_read_ctl would try and read the
  adjacent register, which does not exist, returning crap
* this would in turn lead to wrong size values, e.g., in tapasco-debug
* fixed, is now uint32_t

+++ platform/common/include/platform_info.h
@@ -39,7 +39,7 @@ typedef struct platform_info {
 	} clock;
 	struct {
 		platform_kernel_id_t kernel[PLATFORM_NUM_SLOTS];
-		size_t memory[PLATFORM_NUM_SLOTS];
+		uint32_t memory[PLATFORM_NUM_SLOTS];
 	} composition;
 	struct {
 		platform_ctl_addr_t platform[PLATFORM_NUM_SLOTS];
